--- a/bin/batch.sh
+++ b/bin/batch.sh
@@ -13,4 +13,5 @@ ruby lib/batch/update_statistics_total_downloads.rb
 ruby lib/batch/update_statistics_daily_downloads.rb
 ruby lib/batch/update_trends.rb
 ruby lib/batch/update_gems_latest_columns.rb
+ruby lib/batch/update_daily_summary.rb
 ruby lib/batch/update_master.rb