--- a/lib/models/daily_summary.rb
+++ b/lib/models/daily_summary.rb
@@ -1,0 +1,5 @@
+class DailySummary < Sequel::Model(:daily_summaries)
+  def self.fetch(date)
+    self.where(date: date).first
+  end
+end