… lib/database.rb
… test/models/ranking.rb