Calling `Ranking.total_count` and `Ranking.daily_count` reduce performance of `/gems/:name` page.
Therefore I add `daily_summaries` table to cache these aggregated values.