@xmisao 2017-11-05
Siege benchmark results.
It is 4 times faster.

## Before

Ref: https://github.com/xmisao/bestgems.org/pull/35#issuecomment-338460721

```
$ siege -b -c 4 -r 10 -f bestgems_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 4 concurrent users for battle.
The server is now under siege...
Transactions:                    160 hits
Availability:                 100.00 %
Elapsed time:                   6.99 secs
Data transferred:               4.33 MB
Response time:                  0.17 secs
Transaction rate:              22.89 trans/sec
Throughput:                     0.62 MB/sec
Concurrency:                    3.92
Successful transactions:          80
Failed transactions:               0
Longest transaction:            0.65
Shortest transaction:           0.01
```

## After

```
$ siege -b -c 8 -r 10 -f bestgems_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 8 concurrent users for battle.
The server is now under siege...
Transactions:                    320 hits
Availability:                 100.00 %
Elapsed time:                   3.63 secs
Data transferred:               8.83 MB
Response time:                  0.09 secs
Transaction rate:              88.15 trans/sec
Throughput:                     2.43 MB/sec
Concurrency:                    7.68
Successful transactions:         160
Failed transactions:               0
Longest transaction:            0.30
Shortest transaction:           0.01
```